+++ CHANGELOG.md
@@ -11,6 +11,14 @@ This project adheres to
 
 - `Version.FullString` function that returns the version as a string that
   includes the build metadata.
+- `Version.StrictEqual` function that compares the whole version data to
+  determine the equality of two versions, including build metadata.
+
+### Changed
+
+- `Version.Equal` function to only compare the version parts up to the build
+  metadata as the build metadata is not comparable in the semantic versioning
+  specification.
 
 ### Fixed